# docs/openapi/autoflow-api.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "AutoFlow API",
    "version": "v1",
    "description": "Contrato OpenAPI versionado da API AutoFlow. Exportado a partir do contrato SpringDoc da aplicação e complementado com parâmetros de execução, autenticação e exemplos em docs/openapi/README.md."
  },
  "servers": [
    {
      "url": "http://localhost:8081",
      "description": "API local (backend)"
    },
    {
      "url": "http://localhost:4200/api",
      "description": "API local pelo proxy do frontend"
    }
  ],
...
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
    "/auth/cadastro": {
      "post": {
        "operationId": "cadastrarUsuario",
        "summary": "Cadastrar usuário",
        "tags": [
          "autenticação"
        ],
        "description": "Cadastrar usuário",
        "responses": {
          "201": {
            "description": "Operação concluída com sucesso.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsuarioCadastroResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegistroRequest"
              }
            }
          }
        },
        "security": []
      }
    },
    "/auth/login": {
      "post": {
        "operationId": "login",
        "summary": "Autenticar usuário",
        "tags": [
          "autenticação"
        ],
        "description": "Autenticar usuário",
        "responses": {
          "200": {
            "description": "Operação concluída com sucesso.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              },
              "example": {
                "cpfCnpj": "52998224725",
                "senha": "<senha-do-ambiente>"
              }
            }
          }
        },
        "security": []
      }
    },
    "/auth/usuarios": {
      "get": {
        "operationId": "listarUsuariosAuth",
        "summary": "Listar usuários",
        "tags": [
          "autenticação"
        ],
        "description": "Listar usuários Permissão: ADMIN.",
        "responses": {
          "200": {
            "description": "Usuários retornados com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UsuarioResponse"
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
    "/auth/mecanicos": {
      "get": {
        "operationId": "listarMecanicos",
        "summary": "Listar mecânicos",
        "tags": [
          "autenticação"
        ],
        "description": "Listar mecânicos Permissão: ADMIN ou ATENDENTE.",
        "responses": {
          "200": {
            "description": "Operação concluída com sucesso.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UsuarioResponse"
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
    "/usuarios": {
      "get": {
        "operationId": "listarUsuariosStaff",
        "summary": "Listar usuários pelo staff",
        "tags": [
          "usuários"
        ],
        "description": "Listar usuários pelo staff Permissão: ADMIN ou ATENDENTE.",
        "responses": {
          "200": {
            "description": "Operação concluída com sucesso.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UsuarioResponse"
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      },
      "post": {
        "operationId": "cadastrarUsuarioStaff",
        "summary": "Cadastrar usuário pelo staff",
        "tags": [
          "usuários"
        ],
        "description": "Cadastrar usuário pelo staff Permissão: ADMIN ou ATENDENTE.",
        "responses": {
          "201": {
            "description": "Operação concluída com sucesso.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsuarioResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegistroRequest"
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Use o token retornado por POST /auth/login no header Authorization: Bearer <token>."
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Requisição inválida, JSON inválido, validação ou regra de negócio.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "erro": "JSON inválido"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "JWT ausente/inválido ou token público ausente, inválido ou expirado.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "erro": "Token inválido ou expirado"
            }
          }
        }
      },
      "Forbidden": {
        "description": "Usuário autenticado sem a role necessária.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "erro": "Acesso negado"
            }
          }
        }
      },
      "NotFound": {
        "description": "Recurso não encontrado.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "erro": "Recurso não encontrado"
            }
          }
        }
      },
      "Conflict": {
        "description": "Conflito de unicidade ou recurso já cadastrado.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "erro": "Recurso já cadastrado"
            }
          }
        }
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "erro": {
            "type": "string",
            "example": "Recurso não encontrado"
          },
          "campos": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          }
        }
      },
      "LoginRequest": {
        "type": "object",
        "properties": {
          "cpfCnpj": {
            "type": "string",
            "example": "52998224725"
          },
          "senha": {
            "type": "string",
            "format": "password",
            "example": "<senha-do-ambiente>"
          }
        },
        "required": [
          "cpfCnpj",
          "senha"
        ]
      },
      "LoginResponse": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "example": "<jwt-issued-at-runtime>"
          }
        },
        "required": [
          "token"
        ]
      },
      "RegistroRequest": {
        "type": "object",
        "properties": {
          "nome": {
            "type": "string",
            "example": "Maria da Silva"
          },
          "email": {
            "type": "string",
            "format": "email",
            "example": "maria@example.com"
          },
          "cpfCnpj": {
            "type": "string",
            "example": "12345678901"
          },
          "telefone": {
            "type": "string",
            "example": "11999999999"
          },
          "senha": {
            "type": "string",
            "format": "password",
            "example": "<senha-do-ambiente>"
          },
          "role": {
            "type": "string",
            "enum": [
              "ADMIN",
              "MECANICO",
              "CLIENTE",
              "ATENDENTE"
            ],
            "default": "CLIENTE",
            "example": "CLIENTE"
          }
        },
        "required": [
          "nome",
          "email",
          "senha"
        ]
      },
      "UsuarioResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "example": 1
          },
          "nome": {
            "type": "string",
            "example": "Maria da Silva"
          },
          "email": {
            "type": "string",
            "format": "email",
            "example": "maria@example.com"
          },
          "role": {
            "type": "string",
            "enum": [
              "ADMIN",
              "MECANICO",
              "CLIENTE",
              "ATENDENTE"
            ],
            "example": "CLIENTE"
          }
        }
      },
      "UsuarioCadastroResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "example": 1
          },
          "nome": {
            "type": "string",
            "example": "Maria da Silva"
          },
          "email": {
            "type": "string",
            "format": "email",
            "example": "maria@example.com"
          },
          "role": {
            "type": "string",
            "enum": [
              "ADMIN",
              "MECANICO",
              "CLIENTE",
              "ATENDENTE"
            ],
            "example": "CLIENTE"
          }
        }
      },
...
    }
  }
}